Rearanging project structure before for nuget publishing
