--- mockroservices-dotnet/VaughnVernon.Mockroservices/SourceType.cs
@@ -1,6 +1,0 @@
-﻿namespace VaughnVernon.Mockroservices.VaughnVernon.Mockroservices
-{
-    public interface ISourceType
-    {
-    }
-}
