--- mockroservices-dotnet/packages.config
@@ -1,17 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="JsonNet.PrivateSettersContractResolvers.Source" version="0.1.0" targetFramework="net47" developmentDependency="true" />
-  <package id="Microsoft.Web.Xdt" version="2.1.1" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net45" />
-  <package id="NuGet.Client" version="4.2.0" targetFramework="net45" />
-  <package id="NuGet.Common" version="4.5.0" targetFramework="net45" />
-  <package id="NuGet.ContentModel" version="4.2.0" targetFramework="net45" />
-  <package id="NuGet.Core" version="2.14.0" targetFramework="net45" />
-  <package id="NuGet.Frameworks" version="4.5.0" targetFramework="net45" />
-  <package id="NuGet.Packaging" version="4.5.0" targetFramework="net45" />
-  <package id="NuGet.Packaging.Core" version="4.5.0" targetFramework="net45" />
-  <package id="NuGet.Packaging.Core.Types" version="4.2.0" targetFramework="net45" />
-  <package id="NuGet.Repositories" version="4.2.0" targetFramework="net45" />
-  <package id="NuGet.RuntimeModel" version="4.2.0" targetFramework="net45" />
-  <package id="NuGet.Versioning" version="4.5.0" targetFramework="net45" />
-</packages>
