+++ src/VaughnVernon.Mockroservices.Tests/EventSourcedRootEntityTest.cs
