+++ src/VaughnVernon.Mockroservices.Tests/SerializationTest.cs
