+++ src/VaughnVernon.Mockroservices/InformationExchange.cs
