+++ src/VaughnVernon.Mockroservices/KVStore.cs
