+++ src/VaughnVernon.Mockroservices/Serialization.cs
