+++ src/VaughnVernon.Mockroservices/VaughnVernon.Mockroservices.csproj
@@ -4,13 +4,11 @@
     <TargetFramework>netcoreapp2.1</TargetFramework>
     <RootNamespace>VaughnVernon.Mockroservices</RootNamespace>
   </PropertyGroup>
-
   <ItemGroup>
-    <Compile Remove="Test\**" />
-    <EmbeddedResource Remove="Test\**" />
-    <None Remove="Test\**" />
+    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+      <_Parameter1>$(MSBuildProjectName).Tests</_Parameter1>
+    </AssemblyAttribute>
   </ItemGroup>
-
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
     <PackageReference Include="Microsoft.NETCore.Platforms" Version="2.1.1" />
